requirements.txt
python-dotenv==1.0.1
boto3==1.34.110
botocore==1.34.110
certifi==2024.2.2
charset-normalizer==3.3.2
confluent-kafka==2.4.0
contourpy==1.2.1
Counter==1.0.0
cycler==0.12.1
filelock==3.14.0
fonttools==4.51.0
fsspec==2024.5.0
idna==3.7
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.5
kneed==0.8.5
lapx==0.5.9
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
pika==1.3.2
pillow==10.3.0
psutil==5.9.8
py-cpuinfo==9.0.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-queue-reader==1.2
pytz==2024.1
PyYAML==6.0.1
requests==2.32.2
s3transfer==0.10.1
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
setuptools==70.0.0
six==1.16.0
sympy==1.12.1
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.4
typing_extensions==4.12.0
tzdata==2024.1
ultralytics==8.2.26
urllib3==2.2.1
uugai-python-color-prediction==3.0.0
uugai-python-dynamic-queue==1.1.0
uugai-python-kerberos-vault==1.2.0
